Allow checking exports even if the exports feature is not set.
It is important to check for the export feature when starting a job.

*However*, some exporting functionality does not require a the "exports"
feature to be enabled - such as exporting import data or the species list.

It is nice for exports that are not feature-gated to share the same code
as the feature-gated exports, and checking if an export is does nothing if you
can't start the export to begin with, so this should not interfere with
existing feature-gating.