Convert TreeFavorite into Favorite
The TreeFavorite class was created as part of the OTM1 migrator.
It makes more sense in OTM2 to favorite a Plot or Map Feature, not a Tree.

Because of that, the TreeFavorite model is now just Favorite, and has a foreign
key to MapFeature, not Tree.  All existing TreeFavorites are converted into
Favorites via a data migration.