Fix species export
This is a tricky situation. The call to
`extra_select_and_values_for_model` was being executed with the wrong
model_name ('species' instead of 'Species'), so no field names were
returning. As a result, a call to `qs.values(*[])` was
retrieving *every* field, so the CSV in the unit test had extra fields.

Once the call was corrected, the CSV was being produced correctly, but
another corner case was exposed: Unit tests for CSVs will fail if the
test criteria falls in the first field. This is because the presence of
a BOM in the field header made it a name column lookup by field fail.