Don't localize CSV values
We're already doing this for our exports that use djqscsv, this brings
other exports up to date.

I came across a situation where a test that used to pass began failing
because our suite has gotten large enough to produce primary keys that
receive a comma after localization. This is required to fix these tests.