Remove audit sort by id to increase query speed
When the edits page was created as part of 369749a7, an additional sort
by id was added to the queryset. This technically produces more correct
output, since the audits will be displayed in the exact order they were
written to the table, but it causes a 5x slowdown in query execution
time.