Add updated_at to MapFeature
We have decided to keep a denormalized copy of the time each MapFeature
was last updated to facilitate simple querying. This commit adds a new
updated_at field and functionality to keep it maintained.

Managing updates to this new field when only a child object
changes (Tree, Photo, etc.) is tricky. I opted to target the easily
findable save methods and add a call to a new function that updates the
parent MapFeature row directly, rather than depend on save_with_user
being called on the MapFeature instance when none of the properties of
that object have changed.

Adding updated_at created an interesting field permission wrinkle. I did
not what to have to create explicit permission objects to control this
"special" field. Adding a tree photo will require permission to edit
updated_at, even if the user has no other writable permissions on that
object, so I opted to use a class-level property to flag a field as
being world writable.