Mark export jobs as failed on async exception
fixes #1370 on github

This changeset wraps the logic of async export jobs to mark the job as
failed in the case of an uncaught exception. Jobs will be marked failed
and then the exception will reraise, for logging or further handling in
the future.

If celery is not running, jobs will still poll forever, and once celery
is started, they will be picked up and finished.