Allow map admins to disable exports
This commit adds a boolean attribute to the instance model for
controlling whether or not non-administrators can download data as
CSV.

The flag is checked by a helper method both at the template level, to
hide the export button, and at the task level to prevent a template bug
or manually crafted request from "leaking" the data when the admins have
disallowed it.

The new helper extends the existing check of whether the export feature
is enabled at all and, as a result, applies that check in more places
than it was being applied previously.