Allow resetting of diameter/circumeference fields to undefined.
Previously, clearing a field would change it's value to 0.  Thus, if you
accidentally saved a diameter for a tree when you were unsure of the
diameter, there was no way to undo your action.  This change allows empty
values to pass through, so the backend can clear the field and set it
back to None.

Fixes #978