Wrap disabled buttons in a titled label element
connects #2020 on github

Title (help) text is not shown for disabled buttons. However, we want to
explain why buttons are disabled so the user can take action to enable
them.

The recommended fix for twitter bootstrap tooltips is to wrap disabled
buttons in a div or label. In this case, we are using this technique,
but not activating the bootstrap tooltip. This gives us native title
text instead of tooltip windows. This is advantageous because running
$container.tooltip() is expensive, and with how our code is currently structured,
would have to be done on each map_feature click. The downside is that
bootstrap tooltips look prettier than native title text. This change
could me made easily at a later time.

This commit breaks styling in at least two known places:
* The counts/export/add titlebar is misaligned. Unchecking 'display:
  inline-block' on the label fixes this, but breaks the newly added title
  text.
* The 'Add a Tree' option in the topmost navbar is broken. @jfrankl helped
  me identify that this is because we use '>' selectors in the CSS, which
  don't respond flexibly to heirarchy changes.