Use leaf_subclasses for assigning model perms
This is done in lieu of adding Bioswale to the list

Rather than take all subclasses (effectively "all models") and build a
running list of abstract models below it, we take all non-abstract
models, and maintain a list of all abstract models that are used
directly. Practically, this will likely work better, because the quantity of
abstract models we use directly must grow slower than the total number
of abstract models.