Work around Leaflet afterimage issue.
There is a known Leaflet issue which causes an "afterimage" when going one
zoom level beyond a layer's min/max zoom level.
(https://github.com/Leaflet/Leaflet/issues/1905)

This fix works around the issue.  We should be able to remove this once
we upgrade to Leaflet 1.0, once it is out.

Connects to #2089