Fix saving models when collection UDFs are not writable.
Setting collection UDFs to read-only would prevent editing a model if
that model already had collection UDF data on it.  The fix is rather
simple - we no longer attempt to save collection UDF data if there have
been no changes.

I also fixed an issue with clean_value for collection UDFs where it would
attempt to parse a datetime object as a formatted string and turn it into
a date.  This was mostly a symptom, not a cause of the underlying issue,
but it is more correct anyways.