Raise a 400 when a delete fails
previously, the delete_map_feature endpoint always returned 200. This
means that the map would redirect as if delete succeeded, but the
resource would still be there.

This case would only turn up if a user loads the page, then an admin
reduces their permissions, and then they attempted to delete.