Add additional perm check for deletion
It is not ideal for the template (front-end) to use a different
permission check than the backend. In order to guarantee parity, the
check used for the client is also used on the server. Since these
functions use object_caches under the hood, this will not be costly for performance.