Fix audit unit format for multi-word names
connects #2248 on github

As was the case with bb11caad41f9b2feb1da8c387f9ddef6be5f90aa, unit
field lookup was failing for case-sensitivity reasons on multi-word
models. This is because we were calling `.lower()` on a camel-case
string. In this case, we were producing 'rainbarrel' when we wanted 'rainBarrel'.