Add benefit counts to search bar for Rain Barrel
This was a tricky bug to find, because resource counts worked for
Bioswale and RainGarden, but not RainBarrel. The reason is that these models inherited from
PolygonalMapFeature, which specifically overrides `benefits`
definition. RainBarrel inherits directly from MapFeature which didn't
provide a general definition, until now.

This is largely the result of keeping our object counts married to
our benefit calculations.