Export the inlineEditForm getDataToSave method
This is the "serialization" method for the inlineEditForm. Exporting it
allows access to the same data that would be PUT when the form is
saved, which enables a "preview" workflow. The data can be sent to the
server and the server can return simulated results for verification
before running the actual save operation.