Always pass a falsy predicate to unrelated model
connects #841 on github

The primary purpose of catching udf model parse exceptions instead of
allowing them to raise is to prevent signal crossing between the filter
building of different models. For legacy reasons, udf predicates go to all models, but
they should produce falsy predicates for models that they are not
related to.