Don't check datatype when querying udfc existence
This method will look for a udfc and attempt to create it, but it gates
on *more* fields than a downstream check right before creation. In other
words, existing objects can pass this test and appear not to exist, but
then fail a later test and cause a validation error.