Remove separate TREE_MAPPING for search clauses
Currently when searching, we will take a filter and send it to all subclasses of MapFeature for a search and then merge their results. However, TREE_MAPPING is used for the case of plot, but will fail to find keys for clauses that represent searches on non-plot mfts. By removing TREE_MAPPING and instead deriving the trees from the plots queryset, we remove the need for a separate mapping that doesn't include all the possible clauses to search by.