Handle import parsing of UDF dates and multichoice
When importing a tree CSV with date and/or multichoice UDF columns, data
in an invalid format would trigger the throwing of an `ValueError` which
crashed the celery validation task. In this commit I continue the
pattern of other UDF validations by catching the `ValueError` and
converting it to a `ValidationError`, which is caught and logged by the
validation task.