Handle importing multichoice UDF values as JSON
Multichoice field values are stored as a serialized JSON array. This
commit ensures that JSON embedded in a CSV
column (e.g. "[""Scar"",""Rot""]") will import successfully.

Prior to this commit it was possible to attempt to import values into a
multichoice field, correct the validation errors through the import UI,
but end up adding trees to the map with invalid UDF data. This is fixed
by running all imported multichoice values through `json.loads` when
saving them to the model.

Requiring that the CSV contain properly escaped JSON is not user
friendly. Offering a simpler format is a good idea, but out of the scope
of this commit, which just ensures that the UDF values do not get
corrupted.