Hide panning controls on street view.
I attempted to hide the full screen button but as far as I can tell
there is no option to do so.

There is also no way to minimize the address info box, so I left it in
place but increased the height of the streetview element so it did not
get in the way as much.

Connects to #2364