Refactor itree_codes API
Reduce the number of itree_region methods on instance down to two. Does
this by rolling the use cases `itree_region_codes` and `itree_regions`
together, so that the singular method `itree_regions` produces a rich
enough data structure to be useful for all callers.

Normally, you would expect `itree_regions` to be the source of truth and
`itree_region_codes` to be a helper that reduces `itree_regions` to a
simpler type. You can then imagine a natural flow of query and simplification: get your dataset from
the source of truth, filter it down to relevant records, simplify the
records down to the relevant attributes for your use case. However, since itree region data is sometimes stored as a
string (instance.itree_region_default) and sometimes as a model (ITreeRegion), the source of truth for tying them together
previously had to deal in the lowest common denominator, strings. In
several (but not all) cases, this leads to an awkward flow of query and
simplification: get your dataset from the source of truth, filter it
dwon to relevant records, then enhance it with wrapper functions that
join to data from other sources.

With this refactor, we take advantage of ducktyping to lift string-based
codes to a class that satisfies the same API as the database backed
ITreeRegion. This allows callers not to know or care about the
underlying representation, database backed or string-based. Since
`itree_regions` now always returns rich-types, the flow of
simplification is now always the same: get itree_regions for the
instance, filter down records as needed, simplify attributes as needed.

Other changes:
Moves the itree_region method from tree up to mapfeature, for two
reasons:
1) it's not based on any tree fields, it makes more sense on plot or
above
2) it doesn't rely on plot fields either, so it is suitable for any
mapfeature

Adds a helper to tree to calculate itree code easily.