Simplify and consolidate uses of object_caches
Inline some uses of object_cache functions so that it is easy to see, at
a glance, if it is even possible that a mutability problem can
occur. The closer these statements are together lexically, the easier to
analyze. This was useful while performing this analysis in order to push
necessary uses of deepcopy down to the callers.

Remove some methods that are unused, hardly-used, or generally add a
layer of indirection to the object_caches library. In an ideal world, no public
function will close over object_cache functions. The result would be
that one could search for modules that import object_caches and only
analyze those. These changes take a step closer to that end.

Remove the `fields` method from from UDFDictionary, which is an old
cache of udf_defs that had been converted to a simple wrapper
around object_caches.