Always set importer status to canceled when calling the cancel endpoint
On some occasions we cannot get the Celery task when attempting to call
GroupResult.restore(), which caused attempts to call .revoke() to fail,
which caused the change to the Import status to be rolled back.

We explicitly check the return of .restore() before attempting to call
revoke(), which prevents the transaction rollback from reverting the changed
status.

Connects to OpenTreeMap/otm-addons#991