Add flexible 'remove_map_feature_types' api
Make it easier for callers to specify what should be removed, before
passing control to the instance class.

In other words, some callers have special rules for what to
keep/remove. Rather than have them do the removal themselves, we want to
encapsulate the removal process in one place, so it can be expanded for
new behaviors.