Update Boto to latest v2 version.
Boto3 is now out as well, but according to the docs Boto2 will continue to
receive bugfixes, so I think we are fine staying on Boto2 for now.

According to the changelog there were two potential breaking changes:
 - http://docs.pythonboto.org/en/latest/releasenotes/v2.37.0.html
 - http://docs.pythonboto.org/en/latest/releasenotes/v2.25.0.html

We do not use either API, so we should be fine.  All other changes were
bugfixes or new features.