Update Pillow to the latest version.
According to the release notes, there were many bug fixes between 2.3.1
and 3.1.0, including several security fixes, and there are now additional
supported file types.

The major breaking changes are the removal of several deprecated methods,
and the requirement to use zlip and libjpeg, which were previously
optional dependencies, neither of which are breaking changes for us.

There was also a small breaking change to the rotate method:
expand=True became a required argument to when it is desired for the image
size to change. Previously this was inferred to be true if the rotation
angle was a multiple of 90 degrees.  Image rotation is also slightly less
accurate now, so sometimes image sizes change by a few pixels after rotation.