Make batch moderation selector more specific
The single action selector was including the anchor tag for "View
Details" on the photo moderation page, which should be a plan request
rather than an AJAX request that replaces the content of the page with
the response.

The links on the photo moderation page already had a class attribute set
correctly. I added `class="action"` to the links on the comment
moderation page so that they would still trigger AJAX requests when clicked.