Use ESRI suggest endpoint to provide autocomplete of addresses
This removes our client-side geocoder in favor of our (unused) server-side
geocode endpoint using python-omgeo.

Since the suggest endpoint is using the instance's extent as a viewbox, I
was able to remove the viewbox code from the geocode endpoint, as results
will be matched up with the 'magicKey' provided by ESRI suggestion results.

The automcpletion has been wired into both the location search in the
search bar as well as the location search when adding a tree.

The styling also had be fixed up a bit since we now have typeaheads in
places that used to just have a normal input.

Connects to #948