Fix race condition in celery subtasks
A parent task creates subtasks here, but sends full django objects to
the subtask instead of ids. It is a known convention that one should
send ids to tasks and requery to avoid stale objects, it turns out the
same is true for subtasks. In this case, the parent task would send the
object with a certain value for `.status`, change it and save. The sub
task would then resave `.status` with the stale value.

This is similar in outcome to the previously discovered race condition, but due
to a different cause.