Add more bookkeeping to import event progress
This changeset introduces `last_processed_at` and `is_lost`.

The purpose of `last_processed_at` is to provide an async-safe, guaranteed way of
determining if this import event is being actively processed. Whenever
the event enters an state of active processing, this timestamp will be
repeatedly updated. When it leaves an active processing state, this
timestamp is cleared. As a result, any import even with an old timestamp
in this column must have lost a worker in an unrecoverable failure. It
is safe to clean up this import event.

The purpose of `is_lost` is to add a flag for permanently unrecoverable
jobs without clobbering what is already present in the `status`
column. By having a separate column, all import events with `is_lost` ==
't' will preserve the state in which they failed.