Create InstanceBounds model and migrate data
This changeset (including subsequent commits) detaches bounds, which may be large, complicated geometries, from instances. This means that bounds objects will only be loaded when actually used. This results in a large importer speedup (~60%) for instances with complicated bounds. The speedup should affect most pages on the site, but may only be noticeable on large instances, if at all.

These migrations are meant to gracefully support blue/green
deployments. There will be follow-up migrations in other commits that
will be synchronized with the release cycle.