Stop using instance.bounds in favor of bounds_obj
connects #2399 on github

This is straightforward in most cases: walk the relationship from
Instance to InstanceBounds and access the geom column.

One exception was the way we used the (soon to be deprecated)
`GeoQuerySet.distance()` method, which annotates with distance
field. This method does not ask which field to annotate with, so you cannot specify
something like `instances.distance(point,
annotate_with='bounds_obj__geom'). As an alternative, we iterate over
the instances and annotate ourselves. I have expanded the unit test coverage in this area to confirm it is
working and provide more safety.