Conditionally disable collection UDF search dates
The advanced search UI was allowing the selection of a
date range for stewardship/alerts without selecting a type of either
stewardship or alert. We don't support searching across collection UDF
types, so this resulted in the code generating invalid filter query
strings.

This commit addresses the issue by following the existing patterns in
`udfcSearch.js` and disabling the date fields until both a model and a
CUDF type have been selected.