Update log handling for ecobenefit failures
Failing to calculate the ecobenefits for a tree for reasons like the
species not having a i-Tree code defined for a given i-Tree region is a
regular occurrence in production. This commit attempts to makes it
easier for a log aggregator to combine similar failures and reduces the
severity level of commonplace failures to "info" so they do not drown
out more concerning warning and error messages.

Unexpected failures not related to the species and region configuration
are now logged as errors because we want to see and address those issues
quickly.