Refactor inlineEditForm save API
This commit does a few high level things:
1) decouples onSaveBefore from getDataToSave
2) decouples onSaveAfter from update
3) enhances onSaveBefore to support blocking operations
4) removes the usecase for exporting getDataToSave

This change was prompted by a new feature that needs to conditionally raise a
confirmation modal based on data gather from the form, block on the
modal, and finally continue with the save process.

I first considered adding prompt modals to the inlineEditForm via a
series of hooks, but the process would be too cumbersome. Some modals
have dynamic content, some forms use a series of modals, some interleave
AJAX requests. The alternative and chosen approach, using streams and
flatMapping operations, allows custom workflows of all types as long as
they can finish by returning a stream that fires when it's ok to
continue saving.

In order to make this enhancement to onSaveBefore, it had to be
decoupled from getDataToSave. Otherwise, getDataToSave would have had to
also be modified to support streams in order to pass through
correctly. Fortunately, this has the side-benefit of better separation
of concerns. It was not required to decouple
onSaveAfter, but this was done for the same side benefit.

onSaveBefore now supports block operations via a stream API. This is
documented in the source code and omitted here.

getDataToSave is no longer publicly exported because its use case can be
covered by the new change to the API. This is covered in another PR, but
essentially, it means passing a callback in as onSaveBefore instead of
calling getDataToSave within a graph of jQuery callbacks.