Provide a globalCancelStream to callers
This was probably always correct but its case hadn't turned up as
necessary until now. Callers with advanced uses of the inlineEditForm sometimes need to
trigger a cancel event manually. It should also be possible to listen
for those cancels as part of the exported cancelStream. A caller in an
accompanying PR makes use of this so that it can trigger a cancel from a
modal and then listen on the cancelStream to clean up its UI.

To avoid the daunting task of testing for regressions in
every listener of the cancel stream, this API is left separate with a
TODO.