Increase power of _csv_field_header_map
It is desirable for _csv_field_header_map to be able to make changes to
the included fields and names in the queryset it is connected to. By
using an ordered dict, we can replace future uses of `ordered_fields`
with `field_header_map`s keys, giving `_csv_field_header_map` more
control over what happens later.