Update robots.txt
When robots.txt is empty, search crawlers request every linked page on
the site. Only a small subset of pages on the site have content that
would be useful in a general web search context:

- Map feature detail pages
- Public user profile pages
- Instance custom pages

It is important that the /instance/edits/ page is disallowed because it
is a paginated list of every edit and can result in millions of crawler
requests if the instance is large.

These URLs should be allowed:
/
/map/
/engage/
/analyze/
/mobile/
/pricing/
/instance/features/12345/
/instance/page/About/
/user/person/

These should be disallowed:
/instance/
/instance/map/
/instance/edits/
/instance/management/
/instance/management/a-management-sub-page/