Preserve error panel's current page when resolving an error
We already had a mechanism for this but just hadn't included the current panel and page number when hitting the `update_row` URL

Connects #2526

Testing:
Export a few trees (more than 10) from one instance and import to another instance.
They should all be on the "Errors" panel (lat/lon out of bounds).
Go to page 2 and change X or Y.
It should update and you should remain on page 2.
Extra credit: Fix it with an appropriate lat/lon and see it move to "Ready to Add"