Add label documentation to CONTRIBUTING.md
We standardized labels across all of our OpenTreeMap repositories. This
new text explains the meaning of each label and links to a cross-repo
search of all issues with the `contributing` as a way of directing
contributors to the issues we would like help with and do not require
deep architectural knowledge of the OTM code base.