Fix random_trees command delete option and add delete_plots command
The --delete option to random_trees was incorrectly attempting to delete
all trees from all maps, not just the map associated with the command.

I switched the option to delete via SQL for better performance and so that
it could delete audits and other related objects as well.

I also added a delete_data command, which deletes all trees, plots, map
features and other related data from the given map, leaving mainly only
the import history and customizations in place.

Connects to #1139