Fix infinite loop of browser map state
When panning or zooming the map, we store the map state in the URL.
Conversely, when changing the URL, we pan or zoom the map to match the new URL.

Because we used a fixed precision when storing lat/lng in the URL, but did not
do so for the state object that stored the current zoom/lat/lng, it was
possible for changes to the map state to cause an infinite loop, where the URL
and map state would constantly change but never get in sync with each other.

By serializing and deserializing our internal state object whenever we change
the URL, we ensure that when we later compare the current state object to the
new state to try and detect changes we'll never encounter these loops.

Connects to #2558