Mask fields after checking favorited status
Instance admins can set the permissions on the Plot.id field to
"Invisible." This is not a useful feature, and we intend to replace it
with a better permission system.

For now we can fix the problem of a map feature not showing up as
favorited on the detail page by performing the favorite check before the
fields are masked.