Show canopy information when clicking canopy-enabled maps
For canopy-enabled instances:
* If click hits a feature, include canopy coverage in popup
* If click doesn't hit a feature, fetch HTML showing canopy coverage
* Display canopy coverage as a table of boundaries at click point and their canopy percentages

Also:
* Increase `maxWidth` of popup
* Remove option to pass size of popup (we aren't using it)

Note that double-clicking to zoom a canopy-enabled map shows a canopy popup you probably didn't want to see.
The easy fix is to use `BU.leafletSingleClickStream` instead of `BU.leafletEventStream` to handle
clicks. But the downside is that normal popups are delayed by e.g. 200ms, which seems noticable.
There's probably a way to incur the delay only for clicks that don't hit a feature, but I didn't see
a simple way to implement that.

Connects #2590