Improve default map extent to be based on feature locations
We previously always centered the map on the map center point with a fixed
zoom.
With this change, we instead position the map so that all features added
to the map are visible. If a map has no trees yet, we position the map so
that all of the map's bounds are visible.

Connects to #1155